Percentage changes in economic variables in the USA.

us_change is a quarterly `tsibble` containing percentage changes in quarterly personal consumption expenditure, personal disposable income, production, savings and the unemployment rate for the US, 1970 to 2016.

Format

Time series of class `tsibble`

Source

Federal Reserve Bank of St Louis.

Examples

us_change

#> # A tsibble: 187 x 6 [1Q]
#>    Quarter Consumption Income Production Savings Unemployment
#>      <qtr>       <dbl>  <dbl>      <dbl>   <dbl>        <dbl>
#>  1 1970 Q1       0.616  0.972     -2.45    4.81           0.9
#>  2 1970 Q2       0.460  1.17      -0.552   7.29           0.5
#>  3 1970 Q3       0.877  1.55      -0.359   7.29           0.5
#>  4 1970 Q4      -0.274 -0.255     -2.19    0.985          0.7
#>  5 1971 Q1       1.90   1.99       1.91    3.66          -0.1
#>  6 1971 Q2       0.912  1.45       0.902   6.05          -0.1
#>  7 1971 Q3       0.795  0.532      0.308  -0.446          0.1
#>  8 1971 Q4       1.65   1.16       2.29   -1.53           0  
#>  9 1972 Q1       1.31   0.457      4.15   -4.36          -0.2
#> 10 1972 Q2       1.89   1.02       1.89   -5.05          -0.1
#> # … with 177 more rows